# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "Concesionarios API.",
    "version": "1.0.0"
  },
  "definitions": {
    "Concesionario": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        }
      },
      "required": ["nombre", "direccion"]
    },
    "Coche": {
      "type": "object",
      "properties": {
        "marca": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        }
      },
      "required": ["marca", "modelo"]
    }
  },
  "paths": {
    "/concesionarios": {
      "get": {
        "summary": "Obtener todos los concesionarios",
        "operationId": "getConcesionarios",
        "responses": {
          "200": {
            "description": "Listado de todos los concesionarios.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionario"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo concesionario",
        "operationId": "createConcesionario",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "concesionario",
            "description": "Datos del concesionario a crear",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Concesionario creado exitosamente."
          },
          "400": {
            "description": "Datos inválidos."
          }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Obtener un concesionario por ID",
        "operationId": "getConcesionarioById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID del concesionario"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del concesionario.",
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          },
          "404": {
            "description": "Concesionario no encontrado."
          }
        }
      },
      "put": {
        "summary": "Actualizar un concesionario por ID",
        "operationId": "updateConcesionario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID del concesionario a actualizar"
          },
          {
            "in": "body",
            "name": "concesionario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Concesionario actualizado."
          },
          "404": {
            "description": "Concesionario no encontrado."
          }
        }
      },
      "delete": {
        "summary": "Eliminar un concesionario por ID",
        "operationId": "deleteConcesionario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID del concesionario a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Concesionario eliminado."
          },
          "404": {
            "description": "Concesionario no encontrado."
          }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "Devuelve todos los coches de un concesionario específico",
        "operationId": "getCochesDeConcesionario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID del concesionario"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de coches del concesionario.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Coche"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Añade un nuevo coche a un concesionario",
        "operationId": "addCocheAConcesionario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID del concesionario"
          },
          {
            "in": "body",
            "name": "coche",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coche añadido al concesionario."
          }
        }
      }
    }
  }
}